[Remove Me] Temporarily bump libcudacxx version to 999.000.000.
This allows us to detect whether or not we're building against
the expected libcudacxx headers while testing. Otherwise we may
pull in the installed CTK headers instead of those we're actually
using.

We're simultaneously developing the libcudacxx CMake package while
adding the submodule to Thrust, otherwise this wouldn't be necessary.